feat(newrelic_cloud_aws_integrations): Add other AWS polling integrations.
